requirements.txt
streamlit==1.19.0
pandas
plost
numpy
pillow
google-cloud-firestore
google-auth-oauthlib
firebase_admin
firestore
json
streamlit-authenticator
altair
plotly
base64